backend/requirements.txt
fastapi
uvicorn
openai
langchain
pymongo
python-dotenv
langchain-community
langchain_openai
langchain-core